(dev/core#1615) run_installer - Fix for recently migrated civicrm-setup
Overview
--------

Fix a recent (intra-alpha) regression from https://github.com/civicrm/civicrm-core/pull/16618

The search for `civicrm-setup` goes slightly awry because the migrated copy of civicrm-setup uses
the core's classloader rather than its own classloader.

Steps to Reproduce
------------------

* Create a site with WP and Civi(master).
* Ensure Civi is uninstalled. (ex: `cv core:uninstall -f`)
* In `wp-content/plugins/civicrm/civicrm`, touch the file `.use-civicrm-setup`
* Go to the WP admin console. In the left navbar, choose "CiviCRM Installer".

Before
------

The search for `civicrm-setup` fails to use the code in `./civicrm/setup`
because `civicrm-setup-autoload.php` is not provided or necessary.  It
displays the old UI.

After
-----

It does find the copy of `./civicrm/setup` and displays the new UI.

The absence of `civicrm-setup-autoload.php` is no decisive.

Technical Details
-----------------